+++ server/src/controllers/ingresoController.ts
@@ -3,8 +3,9 @@ import pool from '../database';
 
 class IngresoController {
   public async list(req: Request, res: Response): Promise<void> {
+    const { userId } = req.params;
     try {
-      const ingresos = await pool.query('SELECT * FROM Ingreso');
+      const ingresos = await pool.query('SELECT * FROM Ingreso WHERE IdUsuario = ?', [userId]);
       res.json({ ingresos });
     } catch (err) {
       res.status(500).json({ error: 'Error al obtener los ingresos' });
@@ -13,7 +14,8 @@ class IngresoController {
 
   public async create(req: Request, res: Response): Promise<void> {
     try {
-      await pool.query('INSERT INTO Ingreso SET ?', [req.body]);
+      const { userId, ...gastoData } = req.body; 
+      await pool.query('INSERT INTO Ingreso SET ?', [{ ...gastoData, IdUsuario: userId }]);
       res.json({ message: 'Ingreso guardado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al crear el ingreso' });
@@ -22,8 +24,9 @@ class IngresoController {
 
   public async delete(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId } = req.body;
     try {
-      await pool.query('DELETE FROM Ingreso WHERE IdIngreso = ?', [id]);
+      await pool.query('DELETE FROM Ingreso WHERE IdIngreso = ? AND IdUsuario = ?', [id, userId]);
       res.json({ message: 'El ingreso fue eliminado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al eliminar el ingreso' });
@@ -32,8 +35,9 @@ class IngresoController {
 
   public async update(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId, ...gastoData } = req.body; 
     try {
-      await pool.query('UPDATE Ingreso SET ? WHERE IdIngreso = ?', [req.body, id]);
+      await pool.query('UPDATE Ingreso SET ? WHERE IdIngreso = ? AND IdUsuario = ?', [gastoData, id, userId]);
       res.json({ message: 'El ingreso fue actualizado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al actualizar el ingreso' });
@@ -42,8 +46,9 @@ class IngresoController {
 
   public async getOne(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId } = req.body; 
     try {
-      const ingreso = await pool.query('SELECT * FROM Ingreso WHERE id = ?', [id]);
+      const ingreso = await pool.query('SELECT * FROM Ingreso WHERE id = ? AND IdUsuario = ?', [id, userId]);
       if (ingreso.length > 0) {
         res.json(ingreso[0]);
       } else {
