+++ server/src/controllers/servicioController.ts
@@ -3,8 +3,9 @@ import pool from '../database';
 
 class ServicioController {
   public async list(req: Request, res: Response): Promise<void> {
+    const { userId } = req.params;
     try {
-      const servicios = await pool.query('SELECT * FROM Servicio');
+      const servicios = await pool.query('SELECT * FROM Servicio WHERE IdUsuario = ?', [userId]);
       res.json({ servicios });
     } catch (err) {
       res.status(500).json({ error: 'Error al obtener los servicios' });
@@ -13,7 +14,8 @@ class ServicioController {
 
   public async create(req: Request, res: Response): Promise<void> {
     try {
-      await pool.query('INSERT INTO Servicio SET ?', [req.body]);
+      const { userId, ...gastoData } = req.body;
+      await pool.query('INSERT INTO Servicio SET ?', [{ ...gastoData, IdUsuario: userId }]);
       res.json({ message: 'Servicio guardado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al crear el servicio' });
@@ -22,8 +24,9 @@ class ServicioController {
 
   public async delete(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId } = req.body;
     try {
-      await pool.query('DELETE FROM Servicio WHERE IdServicio = ?', [id]);
+      await pool.query('DELETE FROM Servicio WHERE IdServicio = ? AND IdUsuario = ?', [id, userId]);
       res.json({ message: 'El servicio fue eliminado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al eliminar el servicio' });
@@ -32,8 +35,9 @@ class ServicioController {
 
   public async update(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId } = req.body;
     try {
-      await pool.query('UPDATE Servicio SET ? WHERE IdServicio = ?', [req.body, id]);
+      await pool.query('UPDATE Servicio SET ? WHERE IdServicio = ? AND IdUsuario = ?', [id, userId]);
       res.json({ message: 'El servicio fue actualizado' });
     } catch (err) {
       res.status(500).json({ error: 'Error al actualizar el servicio' });
@@ -42,8 +46,9 @@ class ServicioController {
 
   public async getOne(req: Request, res: Response): Promise<void> {
     const { id } = req.params;
+    const { userId } = req.body; 
     try {
-      const servicio = await pool.query('SELECT * FROM Servicio WHERE id = ?', [id]);
+      const servicio = await pool.query('SELECT * FROM Servicio WHERE id = ? AND IdUsuario = ?', [id, userId]);
       if (servicio.length > 0) {
         res.json(servicio[0]);
       } else {
